// src/features/settings/components/general-settings/types/types.ts
export type Settings = {
  timezone: string;
  locale: string;
  analyticsEnabled: boolean;
};

export type GeneralSettingsValues = {
  websiteName: string;
  email: string;
  mobileNo: string;
  whatsappNumber: string;
  city: string;
  state: string;
  zip: string;
  country: string;
  address: string;
  bankName: string;
  accountHolderName: string;
  accountNumber: string;
  ifscCode: string;
  facebookLink: string;
  instagramLink: string;
  twitterLink: string;
  linkedinLink: string;
  pinterestLink: string;
  youtubeLink: string;
};